[기능] 카테고리 별 가게 목록 조회 - 총 리뷰 개수, 평균 평점 조회 기능 추가

+++ src/main/java/com/backmin/domains/store/dto/StoreInfoAtList.java
@@ -28,7 +28,7 @@ public class StoreInfoAtList {
 
     private boolean isPackage;
 
-    private int averageScore;
+    private double averageScore;
 
     int totalReviewCount;
 
@@ -43,7 +43,7 @@ public class StoreInfoAtList {
             int maxDeliveryTime,
             int deliveryTip,
             boolean isPackage,
-            int averageScore,
+            double averageScore,
             int totalReviewCount,
             List<MenuInfoAtStoreList> bestMenus
     ) {
